+++ index.html
@@ -251,7 +251,7 @@
     </div>
   </template>
 
-  <script src="./js/main.js"></script>
+  <script type="module" src="./js/main.js"></script>
 
 </body>
