+++ js/data.js
@@ -1,0 +1,43 @@
+const AVATAR_RANGE = {
+  MIN: 1,
+  MAX: 8,
+};
+const OFFER_TITLE = 'Wonderful property in Tokyo';
+const OFFER_TYPE = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
+const OFFER_CHECKSET = ['12:00', '13:00', '14:00'];
+const OFFER_FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+const OFFER_DESCRIPTION =
+  'Comfortable real estate in Tokyo. A cozy option for the weary traveler.';
+const OFFER_PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+const LOCATION = {
+  LAT: {
+    MIN: 35.65,
+    MAX: 35.7,
+  },
+  LNG: {
+    MIN: 139.7,
+    MAX: 139.8,
+  },
+};
+
+export {
+  AVATAR_RANGE,
+  OFFER_TITLE,
+  OFFER_TYPE,
+  OFFER_CHECKSET,
+  OFFER_FEATURES,
+  OFFER_DESCRIPTION,
+  OFFER_PHOTOS,
+  LOCATION
+};
