+++ js/utils/get-random-positive-integer.js
@@ -28,3 +28,5 @@ function getRandomPositiveInteger(a, b) {
 }
 
 getRandomPositiveInteger();
+
+export { getRandomPositiveInteger };
